move PaywalledComponent under Users namespace
