+++ app/views/developers/show.html.erb
@@ -18,7 +18,7 @@
               <div class="md:flex md:items-start md:justify-between md:space-x-4 lg:border-b lg:pb-6">
                 <div class="w-full">
                   <h1 class="text-2xl font-bold text-gray-900"><%= @developer.hero %></h1>
-                  <%= render PaywalledComponent.new(user: current_user, paywalled: @developer, size: :small) do %>
+                  <%= render Users::PaywalledComponent.new(user: current_user, paywalled: @developer, size: :small) do %>
                     <div class="mt-2 flex items-center text-sm font-medium text-gray-500">
                       <%= inline_svg_tag "icons/solid/user.svg", class: "flex-shrink-0 mr-1.5 h-5 w-5 text-gray-400" %>
                       <%= @developer.name %>
